+++ Changelog.md
@@ -1,9 +1,14 @@
 # Rest Framework Generic Relations Changelog
 
+## v1.1.0
+
+* Renamed `determine_serializer_for_data` to `get_deserializer_for_data`
+* Renamed `determine_deserializer_for_data` to `get_serializer_for_instance`
+
 ## v1.0.0
 
-* Add support Django 1.8 and 1.9
-* Add support Django Rest Framwork 3
+* Add support for Django 1.8 and 1.9
+* Add support for Django Rest Framework 3
 * Drop support for earlier versions of Django and DRF
 
 ## v0.1.0